Work in progress, piano roll selection.
Added most of the guts needed for moving in piano roll editor.
Moving doesn't update on dragging because the pattern doesn't get
redrawn on a regular update. There's also no current interface from
Selection that allows us to update during the move.

A few methods are copied from DrumPatternEditor; these ought to be
commonised and PianeRollEditor and DrumPatternEditor given a common
ancestor to hold these and other shared functionality.