Keyboard cursor-driven pattern, song and note properties editing (#847)
* Keyboard cursor-driven pattern, song and note properties editing

* Remove unneeded QtDebug include.

* Add braces to unbraced 'if' statements for H2 style

* Remove Instrument Editor checkboxes from focus chain.

* Updated focus behaviour to hide by default

  - Hide the input focus cursor on mouse clicks
  - Input cursor position follows mouse click positions, but without showing the cursor
  - Reveal on tabbing into a widget, or recognised keypress

* Cursor key selection and movement in song editor

* Fix broken ensureVisible call in SongEditor.

* Factor out cursor coordinate conversions in SongEditor

* Standard selection keys (shift+arrows)

* Allow cancellation of select or move with Escape

* Spacing

* Fix spacing, make keyboard input track mouse in drags

* Add continued drawing & toggle in select mode in SongEditor

* Use 'ctrl' in select mode to add or remove from selection

The interaction with dragging is annoyingly fiddly, since Ctrl+drag
starts a move (maybe), so discerning drags from clicks is a pain. Idea
for cleaning this up: create a cleaner separation between UI events
and intentions, to separate out recognition of clicks vs. drags and
Ctrl pressing from their actions.

* Set cursor to copy or move cursor when moving patterns

* Refactor to separate drag detection from actions

* Remove 'override' keywords to avoid consistency warnings.

* Add '+', '-' and Enter actions to NotePropertiesRuler

* Set last value on mouse events

The keyboard cursor behaviour is in general that mouse events will
set the keyboard cursor position as well. In the Note Properties
Ruler, this would correspond to clicking setting the value (y axis,
approximately) for the next keyboard default event (enter / return).

* Don't set a default for enter key on NotePropertiesRuler

* Add select all keyboard action in song editor

* Deselect all key shortcut in Song Editor

* Remove '+' and '-' key actions in song and note editor

The '+' and '-' keys are bound globally to increase / decrease
tempo, so using them in the editors clashes with this.

* Make hiding of keyboard cursor optional (on by default)

* Remove unneeded note property writes

* Spacing and commentary

Added comments of the form '// Key:' to potentially help with
extracting key shortcut lists for documentation later on.