Support opening songs with double click etc (#859)
* Support opening songs with double click etc

Adds support for song and playlist opening and drumkit import from
Explorer/Finder/etc.
  - add positional parameters to command line processing for easier
    Linux and Windows support
  - respond to FileOpen events in MainForm to intercept OS file
    open requsets (necessary for opening at all from Finder on Mac
    OS)

* Catch and defer early FileOpen events at start

* Add document type information to Info.plist

* Handle unsaved changes on Mac OS file open event

* Avoid race condition on FileOpen

MainForm handles opening files via an eventFilter on the
QApplication instance, but MainForm can take a while to
initialise and be ready.  The only thing guaranteed to
receive a FileOpenEvent is the QApplication itself. So this
change subclasses QApplication and implements an event()
method that stores the requested filename so that it can be
later delivered to MainForm when it's ready.

* Fix typo, suppress errors on plist check

* Fix keyboard bug

Using MainForm's event filter on the QApplication instance causes
it to swallow keyboard input for keys it knows about, for all
windows, including dialogs. This structure was partially chose to
avoid subclassing QApplication for the QApplication instance, but
now that turns out to be necessary anyway, it's cleaner to instead
receive the QFileOpenEvent in the QApplication and forward that to
the MainForm explicitly.

* Fix application bundle role to "Editor"