Fix selection infelicity when moving notes.
When moving a note that's superimposed over another identical note,
the actual move could move either one of these. Only one of them
is selected though. To make sure that the user can move the same
selection to again to fine-tune placement etc, this change ensures
that if any found note is in the selection, then that one will be
moved (and so the resulting selection after the move should be
functionally identical.)