remove per track variable in Jack
previously this option has been stored in a variable inherited by the AudioOutput and written once the class instance was created. But the `makeTrackOutput` function did use the underlying option in the Preferences all along. This could potentially lead to errors in the clean up of the audio engine buffers and in the sampling. Thus, this inconsistency has been removed in favor for the Preferences variable.