consistent opening of Songs in GUI
There are quite a number of different ways of how to open a new Song from the GUI. But they seem to have evolved from just copy-pasting some lines of code with a bunch of cleanup calls missing at some occasions. E.g. resetting the undo/redo stack is not done in all instances. This is most probably not intended but a bug.

In order to keep things consistent, easy to maintain, and to miss no instance when integrating NSM specific code, I put all of the code in a set of functions in the HydrogenApp class.