preserve archived drumkit
When a session is archived, the drumkit folder containing all the required samples will be copied in the tar ball. In all subsequent sessions the samples will be loaded from that folder instead of the .hydrogen ffolder in the home. But if the drumkit is changed by the user, the former drumkit folder has to be replaced by a link pointing to the new drumkit. The old one will now just be renamed to `drumkit_old` in order to preserve the samples.