fix OSC command signatures
Quite a lot of OSC commands are registered both to work without an argument and with a float. But none of them does actually use the float argument. I removed the second version since it bloats up the list of available commands while providing no additional benefits.

Also, most commands are registered to expect a float argument, rounding it, and providing an integer to internal functions. I changed their signature in expecting an int because it makes the intention of the command more obvious and the code less complex.