Improved handling of pattern name suffixes
Previously, the method for finding an unused pattern name worked by
appending a numbered suffix to the existing name.  This worked rather
badly if the existing name already had a suffix, as a second suffix
would be appended, resulting in names like "Pattern 3 #1 #1".

This commit fixes this behavior.  Now if the pattern name already has
a " #[number]" suffix, the number gets incremented instead of a new
suffix being appended.