Improve name behavior in the pattern prop.s dialog
Previously, editing pattern names in the pattern properties dialog was
inconsistent with editing them inline in the song editor.  Copying a
pattern always appended the string "#2" to avoid duplicate names.

This commit handles duplicate names in the pattern properties dialog
consistently, by using the same methods from the core as are used
elsewhere.