Fix editing pattern name without change
Previously, if a user edited a pattern name inline, and hit Enter
without changing the name, it would get a " #1" suffix appended, despite
the fact that there would be no name conflict anyway.

This commit fixes this behavior, so that now when editing a pattern's
name inline, its own previous name is ignored when finding an unused
pattern name.