display temporary OSC port
in case the default OSC port is already registered to another client, the OSC server will choose a random port for Hydrogen. Beforehand this port was only once shown at the beginning of the log. Now, it is shown both in the popup window and in the OSC preference tab.

In addition, I changed the red color in Preferences to something more reable.