fix use timeline bpm button in ExportSongDialog
If the Timeline was activated in the GUI but no tempo markers included yet, the use timeline bpm button was deactivated nevertheless.

Now, the checkbox reflects the state of the Timeline button in the GUI and can always be toggled. If there are no tempo markers, toggling this button will just have no effect.