Fix cmake find helper's behaviour with cached variables
FindHelper was finding package paths with find_path/find_library, and then
modifying them afterwards with pkgconf configuration variables if they were
there. That results in the local variables in the first run of cmake being
different from the cached values (which will be used in subsequent runs
of cmake).

There's two potential fixes: either update the cached value once the value
to be used is finally calculated, or post-process the cached variables in
the same way when they're used.

This fixes the annoying error:

```
/usr/include/lrdf.h:8:10: fatal error: raptor.h: No such file or directory
 #include <raptor.h>
          ^~~~~~~~~~
compilation terminated.
```