Prevent duplicate names in pattern prop.s dialog
Previously, the pattern properties dialog didn't change pattern names
that were duplicates of existing pattern names, and only tried to
prevent this by disabling the OK button when the name was a duplicate.
A bug in this behavior allowed multiple patterns to have the same name.

This commit changes this behavior by letting the pattern properties
dialog change the entered name just like the inline pattern name editor.
With this change in place, the OK button is also kept enabled at all
times.  If it is pressed without any field having changed, nothing is
pushed onto the undo stack, to avoid confusion.

Fixes #1083.