WIP: Show a full-height cursor for song position
Previously, Hydrogen had only a small marker on the song editor position
ruler to indicate the playback head position in the song.  This commit
extends the line all the way down the song editor, as discussed in #192.

Still a work-in-progress, as I know of a few minor issues it has still.
For one, the line is offset by half a pixel from the arrow on my HiDPI
monitor.  The arrow is also a different color, which is more obvious now
that the line is longer.  Also, occasionally (and mostly at very high
BPM) the line will get slightly out-of-sync between the ruler and the
song editor.  The cursor's movement is quite jerky too, but that's not
new; ideally we would change the update timer's tick rate based on the
amount of time represented by one pixel on the song editor.

[I confess to having copied and pasted a comment written in Italian, a
language that I do not speak.  As I wrote, work-in-progress.]