Draw tick position without using a pixmap
Since it's just a triangle, and there seem to be some minor issues with
alignment of pixmaps and lines drawn by the QPainter, this commit stops
drawing the tick position indicator from an image.  Now it's drawn from
a QPainterPath instead.  This aligns it properly to the vertical line,
and removes pixelation on HiDPI displays.