remove shortcut bound to "l"
I discovered that there is a shortcut bound to "l" which does exactly the same as pressing the stack mode button in the SongEditorPanel. But it will only work in PATTERN_MODE.

I am not quite sure what was its former intention but since no other button is bound to a shortcut, it will not work in SONG_MODE, and it is not documented, I would suggest to remove it.