remove destructive recording
Fixes #1102. As discussed there "It is a stored in the Preferences class but not written to disk. The drop down menus do not integrate into the pattern editor panel nicely. It is hard to discover and not documented." it will be removed.

In case someone is committed to fix/take care of this, we can revert this commit.