option to disable JACK timebase support
Sometimes (but unfortunately hard to reproduce) the BBT information sent by JACK do some really weird stuff. while the frame number gets incremented as usual, the BBT information shows very large and cyclic jumps e.g. beat 5 => 7 => 12 => 5. This messes up Hydrogen and makes it unusuable.

Up to now I only had this problem with Non-timeline and it went away when pressing the J.MASTER button in the GUI once or twice. But I now had a reproducible instance were it was just messed up with nothing that could be done about it. Curiously, when removing all the timebase support from Non-timeline this bug does still show. So, this is maybe a problem of Jack2 itself?

Anyway, there seems to be things in the JACK timebase ecosystem that to wild and mess up with the transport of Hydrogen. Previously there would be nothing the user could do about but to change to another driver. Hydrogen would always be in slave mode in case there was an external timebase master registered to the JACK server.

I introduced an option in the Preferences to disable the timebase support altogether. This way the basic JACK transport will always be usable.

Also, I updated some shortcuts in the PreferencesDialog