resilient driver loading in Preferences
a couple of weeks ago I updated the options in the PreferencesDialog to ensure consistency in style and spelling. This, among others, lead to capitalizing all acronymns.

When updating Hydrogen from an older version the GUI would not recognized the last audio driver if "Jack", "Alsa" or "Oss" nor the last MIDI driver if "JackMidi" or "CoreMidi" were used. This is itself not a major problem since a single selection of the driver in the PreferencesDialog will solve this conflict. But, nevertheless, this is an avoidable inconvenience.

I added a couple of lines to support the out-dated options too. They are intended to remain only temporary - e.g. during the transition 1.0 => 1.1 - and can be dropped in the release thereafter since most of the user might have happened to open Hydrogen between updates.