Allow different LCDTypes for LCDSpinBox widgets
Previously, the LCDSpinBox widget had the LCDType hardcoded to
LARGE_GRAY, preventing its use for parts of the GUI other than the BPM
widget.

This commit gives its constructor a new parameter allowing the LCDType
to be chosen.  This will enable the user of LCDSpinBox widgets in many
other places in the GUI.