Use slots for button clicked methods in LCDSpinBox
Previously, the LCDSpinBox had two public methods to be called whenever
the up or down button was clicked.  This created a need for a separate
slot to be connected to the button's clicked signal, which in turn
called the appropriate method.  This commit removes this intermediate
step by turning those methods into slots, so that they can be connected
to the clicked signal directly.