Use floating-point limits in LCDSpinBox
Previously, the limits of an LCDSpinBox widget were ints, which seems
limiting given that the widget supports floating-point values.  It also
resulted in some less elegant methods of disabling the upper or lower
limit.

This commit changes the limits to be floats.  This allows the limits to
be disabled by setting them to +/-inf, meaning that no finite values are
treated as special (like we used to do with -1).