fix sanity check in DrumPatternEditor
since the new instrument in DrumPatternEditor::functionDropInstrumentUndoAction is created using `new`, it will always be valid and the check against a nullptr is meaningless and does not check whether the loading succeeded.

I fix the check and added a popup window to make the user aware of the problem. (Just to be save. After fixing the associated issue it should not be triggered at all)