guard against segfault in OSC control
when adjusting the volume, pan etc. of a strip and using path value of 0 or below resp. a number bigger than the number of instruments resulted in a segfault.

In addition, when passing -1 to SELECT_NEXT_PATTERN Hydrogen did segfault as well.

Furthermore, when passing -1 to SELECT_INSTRUMENT Hydrogen will segfault too.