Use a condition variable for logger sleep/wake.
Logger previously used wall clock time to sleep / wake. This had
two(ish) unpleasant side effects.

  1. logger traffic would be excessively bursty, which makes it
     harder to read the log on the console in real time, and
     particularly makes it harder to associate particular log lines
     with user actions.
  2. quitting Hydrogen had to wait for the thread to wake up and
     terminate itself before quitting. This adds on average 0.5s
     to the time taken to quit. Not annoying for most users, but
     probably fairly annoying for developers (hi).
     The sleep period was for some reason only 0.1 seconds on
     Windows, but 1.0 everywhere else. I suspect this was a typo.
     Actually there also seems to have been a pointless extra 1s
     wait (0.1s on Windows) after flushing the log. I think this
     is probably a historical artefact.

It also wasted a very tiny bit of CPU time by waking the thread up
every second to check the queue, but it would be tiny.

This change uses a condition variable instead to block the logger,
so it only wakes up when messages are present, and it does so
immediately. It also wakes up immediately to terminate and flush
buffers.