fix rubber band support
sometime - with not so ideal settings - Rubber Band produces artifacts, like harsh noises or multiple afterimages. As a result the amount of data retrieved from the Rubber Band stretcher can differ dramatically from the expected one, e.g. 11444 instead of 183. This caused Hydrogen to write into memory sectors it did not allocate and resulted in undefined behavior.

With the fix such out-of-buffer writes will be detected and the output buffer will be resized.

I also changed the stretching a little bit. Rubber Band does offer two ways of processing audio: in offline mode and in real-time mode. Both of them have been mixed up in the original code.

Now, whenever Hydrogen is told to apply Rubber Band to the samples on-the-fly on tempo changes during playback (RUB button next to the BPM widget) the real-time mode of Rubber Band will be used, which is faster. Else, the (probably) more accurate offline mode is used.

Fixes #1215 and #828