test storing and validation of drumkit
I enabled the tests checking the saving of the drumkit again and introduced a line handling the cleanup.

I also introduced a test checking whether the saved drumkit is valid / conforms with the drumkit.xsd file. Previously it was only checked whether the drumkit could be loaded. This means a freshly saved drumkit could be invalid but have been successfully loaded via the legacy load.