make storing of empty drumkit valid
since the `Drumkit` constructor is intended to create an empty comtainer for the drumkit components and instruments, I did not add them to the constructor itself. But storing an empty drumkit will now yield a valid drumkit.xml file (with respect to the XSD file) nevertheless. The `Drumkit.save_to()` function will create a default drumkit component and instrument list of an empty one and display a warning.

some of the metadata of the Drumkit will be set in the constructor. At least the name must be empty in order to successfully load it.