Import song in session using aux variable
In order to not touch any essential code parts between the beta and final release, I introduced a auxiliary member variable in the Hydrogen class which allows for the import of songs into an NSM session.

Addresses #1040.

The proper version of the fix will move all the parts for loading a song into the core/main.cpp. This way the GUI does not need to be called asynchronously when setting a song and a number of helper variables can be removed to. I am working on this fix and will make a PR against `development` soon.