load song in main instead of GUI
all essential song loading parts have been moved to the core and the initial song is loaded in the main routine after the core is successfully started. So, no more asynchronous loading of a song in the GUI triggered by the core using a updateSongEvent.

All auxiliary variables and functions introduced to make the previous version of the code work with the NSM support have been removed/cleaned up.