remove mixer font, point size & introduce font size
The Mixer font has been remove from the Preferences all together. Instead, the mixer will use the Application font as well.

The point size set for the Application font wont the stored or used anymore. Instead, a new member in Preferences has been introduced called `FontSize`. This will enable the user to choose between small, normal, and large characters. Which point size correspond to which option is up to us to fine tune.