Fix inaccurate grid-line plotting
Pattern editor grid line plotting used integer X positions with an
integer step. The grid width (ratio of pixels per song tick) is
actually stored as a float, though, and this can acquire some fairly
odd values through successive zooming in and out (in and out use
different ratios). So high grid resolutions and odd zoom levels
could allow significant cumulative error in the grid positioning.

It's hard to notice this in the "usual" use cases of one-par batters,
but for long patterns and subsequent low zoom levels, the cumulative
error could be very noticable!

Since it was part of what caused the confusion, this commit also
changes the name of 'm_nGridWidth' in the Pattern Editor classes
to 'm_fGridWidth' to accurately reflect that it is, actually, a
float.