Fix #1275 - CoreAudio assumes nFrames equivalent to buffer len
Device buffer length is not directly related to render buffer length
(as passed to audioEngine_process()), so don't use this to set the
number of frames to process. Also, don't use it to update the song
timer, for the same reason.

Reviewing the lifetime of the pointers `m_pMainBuffer_{L,R}`, the
lifetimes of these buffer pointers are actually contained entirely
within the lifetime of audioEngine_process(), so these pointers
have been removed and replaced with local calls to getOut_{L,R}()
when the buffers are actually needed.

This then removes the dependency on size of buffers, which solves
the problem of dynamically sized buffers which can occur due to
mismatched device and AudioUnit sample rates.