different highlight when entered and focussed
when entering a widget using the mouse the highlight will now have an alpha value to be less pronounced.

Only when clicked (focussed) the widget will be underlined with the solid color. In addition, the underlining persists when the mouse leaves the widget because its still focussed and still receives all key events issued by the user.