Fix #1275 - CoreAudio assumes nFrames equivalent to buffer len (#1279)
* Fix #1275 - CoreAudio assumes nFrames equivalent to buffer len

Device buffer length is not directly related to render buffer length
(as passed to audioEngine_process()), so don't use this to set the
number of frames to process. Also, don't use it to update the song
timer, for the same reason.

Reviewing the lifetime of the pointers `m_pMainBuffer_{L,R}`, the
lifetimes of these buffer pointers are actually contained entirely
within the lifetime of audioEngine_process(), so these pointers
have been removed and replaced with local calls to getOut_{L,R}()
when the buffers are actually needed.

This then removes the dependency on size of buffers, which solves
the problem of dynamically sized buffers which can occur due to
mismatched device and AudioUnit sample rates.

* Don't expect nFrames == buffer size.

* Remove unneeded buffersize param from LADSPA setup.

The size was unused, and used as a proxy for whether or not the
audio system is active, and should be unnecessary.

* Remove unneeded real time tick calculation for live notes

* No need for jitter correction since tick monotonically increases.

The computed BPM that this was applied to seems to be only used in
the GUI and (incorrectly) in calculating rubberband sample duration,
which only happens on sample load.

* Remove inappropriate re-use of driver buffer size for Rubberband
temp buffer size.