rework preference change handling
instead of storing the former valus in the individual classes and letting them check for changes, the PreferencesDialog will tell them what changed and they will update accordingly.

This approach is less efficient, since any color change and not just the relevant ones will trigger an update of the style sheet. But its less code and complexity.