use English locale in LCDDisplay and LCDSpinBox
since the Rotary and Fader class only allow for "." as separator and it seems over the top right now to support the locale of the user in there, the LCDDisplay must display this separator as well. Else things are too confusing.

To have things consistent, all spin boxes will have the same Locale too.