base AudioEngine on TransportInfo class
Goal of this commit is to remove a large overhead and complexity from the audio drivers and to put a smaller one in the AudioEngine instead.

The TransportInfo (things like current frame, ticksize, bpm, transport status) have been moved into the AudioEngine class and the AudioDrivers will only handle stuff dedicated to the corresponding piece of hardware.

The virtual methods play(), stop(), setBpm(), locate(), and updateTransportInfo() have been moved in the AudioEngine as well.

In the new design the AudioEngine class is derived from the TransportInfo. This was choosen with the intent to make a number of setter function protected and others, as well as all the getters,  public at a later state.