rework AudioEngine and TransportInfo states
Previously the m_State variable of the AudioEngine determined whether the notes should be played back and the frame count incremented - so, effectively, whether playback was rolling. TransportInfo::Status was used to update the state of the AudioEngine during the next run of the process callback.

But there were functions using the TransportInfo::Status information as the audio engine state as well as a lot of instances were other classes directly changed the engine state without considering the current cycle of the process callback.

In this updated version the TransportInfo::Status information was dropped in favor for a member variable AudioEngine::m_nextState which more clearly communicates the intent of changing the state during the next cycle. Also, apart from some hand-selected friend functions, AudioEngine::setState() and AudioEngine::setNextState() have been made private. In order to change the state of the audio engine, all other classes must use either AudioEngine::play() or AudioEngine::stop().