Avoid dereferencing destroyed objects
The savePreferences() method was called from the Preferences
destructor. This would examine windows of the HydrogenApp object,
which is destroyed *before* destroying the Preferences object.

Saving preferences in the destructor feels the wrong thing to do,
so this has been removed and replaced with explicit calls to
savePreferences() before destroying relevant state.