Theme: split off Appearance tab settings
this class - and all its stubclasses - will bundle the settings of the Appearance tab in the PreferencesDialog. All members will be still read from/written to the hydrogen.conf file the way it was done beforehand. But in the future there will be a dedicated import/export concerning just those encapsulated members.