Allow notes to overrun for filter decay
If notes would be sustained by decay on the low-pass filter, they
would previously be cut off as soon as the sample ends, and the
resonant filter state would be ignored, with the note ending sharply.

This change allows the notes to contiue playing until the filter
state is negligible.