Song: fix isModified issues
a lot of instances where parameters of the songs were changed went unnoticed. I fixed this by making all setter methods of Song also marking the song as modified and reviewing other parts changes do occur.

In a number of cases I moved the logic of a song being modified from the GUI into the core.