SongEditorPositionRuler: rework bpm widget and its handling
- ensure that text of the tempo marker is readable when using zoom out
- highlight hovered tempo marker and display it on top of neighboring and possibly overlapping markers
- use highlighting color for tag labels "T"
- highlight tempo marker for which the BPM widget was created (in order to edit it)

Previously arbitrary values could be entered in the BPM widget and where to click on the Timeline only affected the default values of the line edits. Now, clicking at an empty column creates a tempo marker and clicking an existing one allows to move, edit, or delete it. Also, there are no line edits anymore but spin boxes.

the nBar field of TempoMarker struct was renamed to nColumn to match the overall naming in the code