first draft of rework transport processing
Previously relocations (in frames) as well as checking for bpm changes have been done in the beginning of the audioEngine_process cycle. But those lines updating the current column, the tick position within the current pattern, and the starting tick of the current pattern (all of them are used throughout various places in the core and GUI) are only reached when transport is rolling. If stopped, the transport position easily got into a bad state whenever the transport was stopped.

In this redesign the update of all the above variables will be done in a single place. After each complete run of the process cycle (transport is rolling) there will be a full update of all of them. In case of a relocation, all are updated as well.

Tempo changes are still checked for in the beginning of the process cycle. But know this is always done using a clean transport state.