second draft of rework transport processing
TransportInfo::m_nFrames do now represent the frames passed since the beginning of the song (previously, this was not true after the first tempo marker was passed). The whole AudioEngine was designed to only work with a constant tickSize throughout the whole song and not to support tempo changes within.

ticks and frames are now decoupled the AudioEngine::computeFrameFromTick or AudioEngine::computeTickFromFrame have to be used to convert between the two (without the Timeline activated or a single TempoMarker in the beginning of the song things does not change internally). This was done in order to ensure the first point.

The AudioEngine and the Sampler are (almost) the only places that do care about frames. All relocating, note handling etc. is done using ticks.

Playback of the Notes does not work properly right now. But this is only an intermediate state anyway.