third draft: double ticks and unit testing
The data type of the ticks has been changed to double within the AudioEngine to guarantuee (and ease) the one to one mapping of frames and ticks. When everything is done this will also allow for new features, like tuplets.

Unit testing is in place for transport processing and relocation. Also, it is assured that all tick intervals processed by updateNoteQueue are aligned.

Note placement and double handling of ticks therein is still to be done