slight alteration of BPM handling concept
The Timeline does no longer update the tempo of the special TempoMarker if the Song tempo was changed via OSC, MIDI, tap tempo or beatcounter. Instead, the Timeline::activate() function is used to store a copy of the tempo used right in the moment the Timeline was activated and this one is used for the special one.

This changes had been applied since every BPM change would have to trigger the recalculation of all frame-based variables.

Adding/deleting a TempoMarker is now properly handled by the AudioEngine and Sampler without the transport state getting out of sync.