SongEditorPositionRuler: fix playhead position
The position of the playhead is now based on the actual transport position of the AudioEngine. Previously, an additional factor relying on the last AudioEngine::m_pPlayingPatterns processed in AudioEngine::updateNoteQueue() was included. But since the latter also includes a lookahead and is ahead of the transport position, the pos of the playhead is glitching