DiskWriterDriver: write note fade out
I am not quite sure how this was not spotted or reported earlier but H2 did chop off every audio signal when reaching the end of the song. For short samples, like toms, placed a couple of ticks before the end this might be okay. But for longer ones, like crash, this is just bad.

This fix is designed in such a way that when merging with https://github.com/hydrogen-music/hydrogen/pull/1403 the DiskWriterDriver will also wait till all FX processing is done.