AudioEngine: review float precision in Note placement
If we require to change precision of the note placement (Note::get_position()) to float at some point, there are not many things we need to change (under the assumption that pattern ticks remain integer)

Two things have to be changed in the AudioEngine:
1. make FOREACH_NOTE_CST_IT_BOUND loop over all notes `(_it)->first >= (_bound) && (_it)->first < (_bound + 1)`
2. add remainder of pNote->get_position() % 1 when setting nnTick as new position.

Sampler::setPlayingNotelength() has to be reviewed as well. But as long as note placement during recording still works with integer positions, this should be fine too.