fix AudioEngine state consistency in looped transport
when looped mode is activated and the transport was looped at least once adding a new pattern to the end of the song or other stuff increasing the song length would lead to a glitch in the transport position and audio playback. (bug was not introduced in this PR but already present)

this has now be both fixed and turned into a unit test.