Song: introducing LoopMode
Previously, whenever the transport was looped at least once, pressing the Loop button in the PlayerControl resulted in an immediate stop of the playback.

Now, using an additional layer of an enum variable Song::m_loopMode, transport will keep rolling till the end of the song. (This bug was already present prior to the PR)