tests: fix export of fade out
I did not took into account that the zero padding in the output channels does not only occur when the Sampler is done rendering notes but that the samples of the notes can be very well be padded with zeros too. While this is not a problem in general and all actual sound data got written, the number of zeros at the end of the audio file did differ depending on the buffer size. This caused reference test file to not be frame-accurate since a different buffer size was set in the config I used when creating the files with respect to the one used in the CI.

Now, all frames passed the end of the song will be checked for zeros and as soon as there are X, with X = 200 for now, consecutive zeros in both channels the song export will be stopped.