Follow playing patterns in Song Editor (#1437)
* Track playing patterns in Song Editor

Songs with many patterns may not fit in the current viewport of the
Song Editor. Depending on how the song is structured, as the viewport
scrolls to show different times, the playing patterns may end up
being off-screen. It would be ideal to be able to follow the
progression of the song in a meaningful and useful way, but since
multiple patterns may be active at any one time, it's non-trivial
to define a useful behaviour that captures more than the simple
case of a song with one pattern active at a time.

As an attempt to define a useful behaviour which captures what the
user might expect to happen, we define the behaviour as follows:
  * If there are no currently playing patterns which are entirely
    visible:
      * Find the position with the smallest amount of scrolling
        from the current location which:
	  * Fits the maximum number of currently playing patterns
	    in view at the same time.

This covers the trivial cases where only a single pattern is playing,
and gives some intuitive behaviour for songs containing multiple
playing patterns where the general progression is diagonal but with
constant (or near-constant) background elements, and the "minimum
scrolling" allows the user to hint if we stray off the path.

* Avoid inappropriately scrolling SongEditor to current pattern

* Implement pattern-follows-song mode

Enabled via preferences, pattern-follows-song will change the
selected pattern to follow the "current" pattern in song mode:

If the currently selected pattern is not one of the currently-playing
patterns, then the first currently-playing pattern will be selected.

* Prefer visible pattern when selecting a new current pattern