Rewrite Timeline handling
Previously there was a single `Timeline` instance in the `Hydrogen` class. Everytime a `Song` was read (not set!!) all tags and tempo markers were removed from the Timeline and the ones stored in .h2song added. It is an edge case but when loading two Songs and setting the first one the Timeline will be in a dirty state.

Now, every song holds the Timeline read from the corresponding .h2song file as a shared pointer and the currently loaded song will set Hydrogen::m_pTimeline. This way the Timeline should always correspond to the associated Song.