streamline handling of song and tempo changes
all checking of and compensating for changes in tick size, tempo, timeline, and song size is now done in dedicated functions of the AudioEngine and the Sampler called by methods introducing the corresponding changes. The functions themselves, especially the ones in the AudioEngine, are less efficient then incorporating the code in the processing of AudioEngine::processPlayNotes and Sampler::renderNote. But since they are rarely called during playback - like when toggling the Timeline, adding/deleteing tempo markers, setting a different tempo, changing the length of a pattern, or toggling a grid cell in the song editor - it might be slightly more efficient in total.

- the computation of the note start has been moved to the Note class to ensure Note::m_nNoteStart and Note::m_fUsedTickSize do not get out of sync
- note starts computations were now called in AudioEngine::updateNoteQueue in order to make AudioEngine::processPlayNotes and the functions in the Sampler less complex as they knwo can be assumed to always work on clean notes