Use pixmap to cache PatternEditorRuler background
The pattern editor ruler already had a background pixmap, but it
was previously only used for a blank background fill. Added everything
except the mobile tick mark to this and added relevant updates and
device pixelratio sensitivity.

This functionality is ripe for refactoring into some sort of
CachedBackgroundWidget or similar.