add unit test for SSL support
Since Qt does not ship its own OpenSSL library, we provide the required .dll in the Windows release for Hydrogen >1.1.1 .

A member of the Qt team stated in a forum that one has to ensure the exact version of OpenSSL is shipped that was also used to compile Qt5. As they might update the used version in future Qt releases we have to check whether the lib provided by us is the correct one.