rework empty song creation
Previously whenever a new song was loaded, the DefaultSong.h2song shipped with Hydrogen was opened. That was not ideal, because a) it is based on a possibly outdated version of the default Song parameters and GMRockKit settings therein and b) it is bound to a specific file (which on top of that also resides in a read-only location)

Now, an empty song is created from scratch within Hydrogen and the reworked Song::loadDrumkit() is used to load a fresh GMRockKit into it.