Restore changes to empty songs
If the last song of the previous session was an empty one, Hydrogen will now attempt to recover changes. This is done using autosave files, which persists until the user either uses "Save As", discards the chages using the QMessageBox, or discards them by hitting "New Song" again.

To accomplish this, an empty song is now handled differently within the GUI. Core and CLI are agnostic of this complexity and will treat it as a usual song.